Added functionality to receive image from gallery
